Distributed code to cpp and hpp files
